ai-trading-bot/requirements.txt
anthropic>=0.40.0
dhanhq>=2.0.0
pyotp>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14
# TA-Lib>=0.4.28  # Requires C library; install separately if needed
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
APScheduler>=3.10.0
python-telegram-bot>=20.0
pyyaml>=6.0
python-dotenv>=1.0.0
filelock>=3.12.0
streamlit>=1.30.0
yfinance>=0.2.30
